Main elxan

+++ src/Services/Job/JobCompany/JobCompany.Business/Dtos/ExamDtos/GetExamIntroDto.cs
@@ -10,8 +10,8 @@ namespace JobCompany.Business.Dtos.ExamDtos
     {
         public string CompanyName { get; set; }
         public int QuestionCount { get; set; }
-        public byte? Duration { get; set; }
-        public decimal LimitRate { get; set; }
+        public short? Duration { get; set; }
+        public float LimitRate { get; set; }
         public string FullName { get; set; }
         public string IntroDescription { get; set; }
     }
