+++ src/Services/Job/JobCompany/JobCompany.Business/Dtos/ExamDtos/GetExamQuestionsDetailDto.cs
@@ -5,7 +5,7 @@ namespace JobCompany.Business.Dtos.ExamDtos;
 public class GetExamQuestionsDetailDto
 {
     public int TotalQuestions { get; set; }
-    public decimal LimitRate { get; set; }
-    public byte? Duration { get; set; }
+    public float LimitRate { get; set; }
+    public short? Duration { get; set; }
     public List<QuestionPublicDto> Questions { get; set; }
 }
