+++ src/Services/Job/JobCompany/JobCompany.Business/Dtos/ExamDtos/SubmitExamResultDto.cs
@@ -4,6 +4,6 @@ public class SubmitExamResultDto
 {
     public byte TrueAnswerCount { get; set; } // Number of correct answers submitted by the user
     public byte FalseAnswerCount { get; set; } // Number of incorrect answers submitted by the user
-    public decimal ResultRate { get; set; } // The success rate of the user
+    public float ResultRate { get; set; } // The success rate of the user
     public bool IsPassed { get; set; } // Indicates whether the user passed the exam
 }
